# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Textile's HTTP REST API Documentation",
        "title": "Textile REST API",
        "termsOfService": "https://github.com/b582q9/go-textile-sapien/blob/master/TERMS",
        "contact": {
            "name": "Textile",
            "url": "https://textile.io/",
            "email": "contact@textile.io"
        },
        "license": {
            "name": "MIT License",
            "url": "https://github.com/b582q9/go-textile-sapien/blob/master/LICENSE"
        },
        "version": "0"
    },
    "basePath": "/api/v0",
    "paths": {
...
        "/threads": {
            "get": {
                "description": "Lists all local threads, returning a ThreadList object",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "threads"
                ],
                "summary": "Lists info on all threads",
                "responses": {
                    "200": {
                        "description": "threads",
                        "schema": {
                            "$ref": "#/definitions/pb.ThreadList"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Adds a new Thread with given name, type, and sharing and whitelist options, returning\na Thread object",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "threads"
                ],
                "summary": "Adds and joins a new thread",
                "parameters": [
                    {
                        "type": "string",
                        "description": "name",
                        "name": "X-Textile-Args",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "type=private,sharing=not_shared,whitelist=",
                        "description": "key: A locally unique key used by an app to identify this thread on recovery, schema: Existing Thread Schema IPFS CID, type: Set the thread type to one of 'private', 'read_only', 'public', or 'open', sharing: Set the thread sharing style to one of 'not_shared','invite_only', or 'shared', whitelist: An array of contact addresses. When supplied, the thread will not allow additional peers beyond those in array, useful for 1-1 chat/file sharing",
                        "name": "X-Textile-Opts",
                        "in": "header"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "thread",
                        "schema": {
                            "$ref": "#/definitions/pb.Thread"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/threads/{id}": {
            "get": {
                "description": "Gets and displays info about a thread",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "threads"
                ],
                "summary": "Gets a thread",
                "parameters": [
                    {
                        "type": "string",
                        "description": "thread id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "thread",
                        "schema": {
                            "$ref": "#/definitions/pb.Thread"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "description": "Adds or updates a thread directly, usually from a backup",
                "tags": [
                    "threads"
                ],
                "summary": "Add or update a thread directly",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "thread",
                        "name": "thread",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/pb.Thread"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Abandons a thread, and if no one else is participating, then the thread dissipates.",
                "tags": [
                    "threads"
                ],
                "summary": "Abandons a thread.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "thread id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/threads/{id}/name": {
            "put": {
                "description": "Renames a thread. Only initiators can rename a thread.",
                "tags": [
                    "threads"
                ],
                "summary": "Rename a thread",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "name",
                        "name": "X-Textile-Args",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/threads/{id}/peers": {
            "get": {
                "description": "Lists all peers in a thread",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "threads"
                ],
                "summary": "List all thread peers",
                "parameters": [
                    {
                        "type": "string",
                        "description": "thread id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "contacts",
                        "schema": {
                            "$ref": "#/definitions/pb.ContactList"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "pb.Block": {
            "type": "object",
            "properties": {
                "attempts": {
                    "type": "integer"
                },
                "author": {
                    "type": "string"
                },
                "body": {
                    "type": "string"
                },
                "data": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "parents": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "status": {
                    "type": "integer"
                },
                "target": {
                    "type": "string"
                },
                "thread": {
                    "type": "string"
                },
                "type": {
                    "type": "integer"
                },
                "user": {
                    "description": "view info",
                    "type": "object",
                    "$ref": "#/definitions/pb.User"
                }
            }
        },
...
        "pb.Cafe": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "api": {
                    "type": "string"
                },
                "node": {
                    "type": "string"
                },
                "peer": {
                    "type": "string"
                },
                "protocol": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
...
        "pb.Contact": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "avatar": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "peers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/pb.Peer"
                    }
                },
                "threads": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "pb.ContactList": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/pb.Contact"
                    }
                }
            }
        },
...
        "pb.Node": {
            "type": "object",
            "properties": {
                "json_schema": {
                    "type": "string"
                },
                "links": {
                    "type": "object"
                },
                "mill": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "opts": {
                    "type": "object"
                },
                "pin": {
                    "type": "boolean"
                },
                "plaintext": {
                    "type": "boolean"
                }
            }
        },
...
        "pb.Peer": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "avatar": {
                    "type": "string"
                },
                "created": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "inboxes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/pb.Cafe"
                    }
                },
                "name": {
                    "type": "string"
                },
                "updated": {
                    "type": "string"
                }
            }
        },
...
        "pb.Thread": {
            "type": "object",
            "properties": {
                "block_count": {
                    "type": "integer"
                },
                "head": {
                    "type": "string"
                },
                "head_blocks": {
                    "description": "view info",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/pb.Block"
                    }
                },
                "id": {
                    "type": "string"
                },
                "initiator": {
                    "type": "string"
                },
                "key": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "peer_count": {
                    "type": "integer"
                },
                "schema": {
                    "type": "string"
                },
                "schema_node": {
                    "type": "object",
                    "$ref": "#/definitions/pb.Node"
                },
                "sharing": {
                    "type": "integer"
                },
                "sk": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "state": {
                    "description": "Deprecated: Do not use.",
                    "type": "integer"
                },
                "type": {
                    "type": "integer"
                },
                "whitelist": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "pb.ThreadList": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/pb.Thread"
                    }
                }
            }
        },
        "pb.User": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "avatar": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    }
}